+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - automatically recompile style
 - loading eager option for images across all elements
 - transition border option to timeline element
+- before/after labels to image comparison element
 
 ### Changed